# backend/doc/api-docs.yml
swagger: "2.0"
info:
  title: PersonalMediaVault Backend API
  description: API to access PersonalMediaVault from a web client.
  version: 2.0.0

# cSpell:ignore allof, anyof, noneof, gyan
schemes:
  - http
  - https
securityDefinitions:
  SessionTokenHeader:
    type: apiKey
    in: header
    name: x-session-token
...
paths:
...
  "/api/media/{id}":
    get:
      summary: Gets media
      description: Gets media asset metadata
      tags:
        - media
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: number
          format: int64
          description: Media ID
      security:
        - SessionTokenHeader: []
      responses:
        401:
          description: Unauthorized (No valid session)
        404:
          description: Not found
        200:
          description: Ok
          schema:
            properties:
              id:
                type: number
                format: int64
                description: Media ID
              type:
                type: number
                format: int32
                description: >
                  Media Type:
                    * 1 - Image
                    * 2 - Video
                    * 3 - Audio
                enum:
                  - 1
                  - 2
                  - 3
              title:
                type: string
                description: "Media title"
                example: "Media title"
              description:
                type: string
                description: "Media description"
                example: "Media description"
              tags:
                type: array
                description: Associated tags
                items:
                  type: number
                  format: int64
                  description: Tag ID
              upload_time:
                type: number
                format: int64
                description: "Upload timestamp (server time, Unix milliseconds)"
              thumbnail:
                type: string
                description: "Path to reach the thumbnail. Empty if no thumbnail. Thumbnails are squared images."
                example: "/assets/b/0/0/thumbnail.jpg"
              duration:
                type: number
                format: double
                description: Duration of the media in seconds (videos and audios)
              width:
                type: number
                format: int32
                description: Image/Video width
              height:
                type: number
                format: int32
                description: Image/Video height
              fps:
                type: number
                format: int32
                description: Frames per second
              ready:
                type: boolean
                description: "Set to true if the media asset finish uploading"
              ready_p:
                type: number
                format: int32
                description: "Number from 0 to 100 in case ready = false to indicate the storage progress"
              encoded:
                type: boolean
                description: "Set to true if the media asset is encoded and ready to play."
              task:
                type: number
                format: int64
                description: "Task ID in case encoded = false"
              url:
                type: string
                description: "URL of the asset to be displayed. Blank if not encoded yet."
                example: "/assets/b/0/0/video.mp4"
              video_previews:
                type: string
                description: "URL of the video previews (frames each X seconds). Replace {INDEX} with a number starting from 0 for the first one."
                example: "/assets/p/0/0/preview_{INDEX}.jpg"
              video_previews_interval:
                type: number
                format: double
                description: Interval for each video preview, in seconds.
              resolutions:
                type: array
                description: Extra resolutions for slow connections or small screens.
                items:
                  properties:
                    width:
                      type: number
                      format: int32
                      description: Image/Video width
                    height:
                      type: number
                      format: int32
                      description: Image/Video height
                    fps:
                      type: number
                      format: int32
                      description: Frames per second
                    ready:
                      type: boolean
                      description: "If set to true, the resolution is encoded and ready for playback"
                    task:
                      type: number
                      format: int64
                      description: "Encoding task ID if ready = false"
                    url:
                      type: string
                      description: "Playback URL"
                      example: "/assets/b/0/0/video_720p.mp4"
              subtitles:
                type: array
                description: Available subtitles for media
                items:
                  properties:
                    id:
                      type: string
                      description: "Subtitles ID"
                      example: "es"
                    name:
                      type: string
                      description: "Subtitles name"
                      example: "Spanish"
                    url:
                      type: string
                      description: "SRT file URL"
                      example: "/assets/b/0/0/subrip.srt"
              audios:
                type: array
                description: Available extra audios (for videos)
                items:
                  properties:
                    id:
                      type: string
                      description: "Audio track ID"
                      example: "es"
                    name:
                      type: string
                      description: "Audio track name"
                      example: "Spanish"
                    url:
                      type: string
                      description: "Audio file URL"
                      example: "/assets/b/0/0/audio.mp3"
              attachments:
                type: array
                description: Available attachments
                items:
                  properties:
                    id:
                      type: number
                      format: int64
                      description: "Attachment ID"
                    name:
                      type: string
                      description: "Attachment file name"
                      example: "example.txt"
                    size:
                      type: number
                      format: int64
                      description: "Size of the attachment in bytes"
                    url:
                      type: string
                      description: "Attachment file URL"
                      example: "/assets/b/0/0/example.txt"
              force_start_beginning:
                type: boolean
                description: "True to indicate clients that this media cannot save it's time status (for songs / short animations)"
              is_anim:
                type: boolean
                description: "True to indicate the video is an animation, for example an animated gif"
              img_notes:
                type: boolean
                description: "True if the image has notes"
              img_notes_url:
                type: string
                description: "URL to fetch the image notes from"
                example: "/assets/b/0/0/notes.json"
              ext_desc_url:
                type: string
                description: "URL to fetch the extended description from. Empty if no extended description."
                example: "/assets/b/0/0/ext_desc.txt"
              time_slices:
                type: array
                description: "Timeline splits, for videos and audios"
                items:
                  properties:
                    time:
                      type: number
                      format: double
                      description: "Start of the time slice, in seconds"
                    name:
                      type: string
                      description: "Name of the time slice"
...
  "/api/media/{id}/subtitles/set":
    post:
      summary: Sets subtitles
      description: Add a subtitles file for videos and audios
      tags:
        - media
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: number
          format: int64
          description: Media ID
        - in: query
          name: id
          type: string
          required: true
          description: "ID for the subtitles. Example: es"
        - in: query
          name: name
          type: string
          required: true
          description: "Name for the subtitles. Example: Spanish"
        - in: formData
          name: file
          type: file
          description: The file to upload.
      security:
        - SessionTokenHeader: []
      responses:
        400:
          description: Bad request.
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * INVALID_SRT - Invalid srt file provided
                    * INVALID_ID - Invalid subtitles ID
                    * INVALID_NAME - Invalid subtitles name
                    * NOT_SUPPORTED - This API only applies to videos and audios
                enum:
                  - INVALID_SRT
                  - INVALID_ID
                  - INVALID_NAME
                  - NOT_SUPPORTED
              message:
                type: string
                description: "Error message"
                example: "Error message"
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
        404:
          description: Not found
        413:
          description: SRT file too large (Max 10 MB)
        200:
          description: Success
          schema:
            properties:
              id:
                type: string
                description: "Subtitles ID"
                example: "es"
              name:
                type: string
                description: "Subtitles name"
                example: "Spanish"
              url:
                type: string
                description: "SRT file URL"
                example: "/assets/b/0/0/subrip.srt"
...
  "/api/media/{id}/audios/set":
    post:
      summary: Sets audio track
      description: Add an extra audio track to a video
      tags:
        - media
      consumes:
        - multipart/form-data
      produces:
        - application/json
      parameters:
        - in: path
          name: id
          required: true
          type: number
          format: int64
          description: Media ID
        - in: query
          name: id
          type: string
          required: true
          description: "ID for the audio track. Example: es"
        - in: query
          name: name
          type: string
          required: true
          description: "Name for the audio track. Example: Spanish"
        - in: formData
          name: file
          type: file
          description: The file to upload.
      security:
        - SessionTokenHeader: []
      responses:
        400:
          description: Bad request.
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * INVALID_AUDIO - Invalid audio file provided
                    * INVALID_ID - Invalid track ID
                    * INVALID_NAME - Invalid track name
                    * NOT_SUPPORTED - This API only applies to videos
                enum:
                  - INVALID_AUDIO
                  - INVALID_ID
                  - INVALID_NAME
                  - NOT_SUPPORTED
              message:
                type: string
                description: "Error message"
                example: "Error message"
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
        404:
          description: Not found
        200:
          description: Success
          schema:
            properties:
              id:
                type: string
                description: "Audio track ID"
                example: "es"
              name:
                type: string
                description: "Audio track name"
                example: "Spanish"
              url:
                type: string
                description: "Audio file URL"
                example: "/assets/b/0/0/audio.mp3"
...
  "/api/albums/{id}/set":
    post:
      summary: Set album list
      description: Sets album media list
      tags:
        - albums
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: path
          required: true
          type: number
          format: int64
          name: id
          description: Album list
        - in: body
          name: Album name
          schema:
            properties:
              list:
                type: array
                description: List of media Ids, in order
                items:
                  type: number
                  format: int64
                  description: Media ID
      security:
        - SessionTokenHeader: []
      responses:
        400:
          description: Bad request.
          schema:
            properties:
              code:
                type: string
                description: >
                  Error code:
                    * INVALID_LIST - Invalid album list
                    * MAX_SIZE_REACHED - The list has more than 1024 elements
                enum:
                  - INVALID_LIST
                  - MAX_SIZE_REACHED
              message:
                type: string
                description: "Error message"
                example: "Error message"
        401:
          description: Unauthorized (No valid session)
        403:
          description: Access denied
        404:
          description: Not found
        200:
          description: Success
...